+++ CHANGELOG.md
@@ -1,4 +1,18 @@
 
+## [v0.6.0] - 2025-10-07
+
+### Changed
+
+- Fixed an issue with streaming long output from C2 profile debugging
+- Added `semver` fields to all service definitions
+- C2 Profiles can now report back agent_icon and dark_mode_agent_icon just like payload types
+- Payload type build parameters now have a supported_os, group_name, and hide_conditions fields
+  - hide conditions allow you to specify when a specific build parameter should be hidden from user view
+  - group_name allows you to group like-parameters together in the UI
+  - supported_os allows you to limit build parameters to certain OS selections
+- Payload type how has a c2 deviations parameter that allows you to modify fields of supported C2 profiles
+  - for example - alter the defaults, change dropdown options, hide parameters completely
+
 ## [v0.5.32] - 2025-05-30
 
 ### Changed